Find location of gdb and DeveloperDiskImage.dmg based on xcode-select

+++ fruitstrap.c
@@ -5,6 +5,7 @@
 #include <sys/socket.h>
 #include <sys/un.h>
 #include <stdio.h>
+#include <dirent.h>
 #include <signal.h>
 #include <getopt.h>
 #include "MobileDevice.h"
@@ -138,6 +139,62 @@ CFStringRef copy_developer_disk_image_path(AMDeviceRef device) {
     if (!found) {
         path = CFStringCreateWithFormat(NULL, NULL, CFSTR("/Developer/Platforms/iPhoneOS.platform/DeviceSupport/Latest/DeveloperDiskImage.dmg"));
         found = path_exists(path);
+    }
+    if (!found) {
+        FILE* select_path = popen("xcode-select --print-path", "r");
+        if (select_path)
+        {
+            char buffer[1024];
+            char* result_path = fgets(buffer, sizeof(buffer), select_path);
+
+            char raw_version[10];
+            CFStringGetCString(version, raw_version, sizeof(raw_version), kCFStringEncodingASCII);
+
+            /* Remove subversion number */
+            bool dotFound = false;
+            int version_index = 0;
+            for(version_index = 0; version_index < sizeof(raw_version); version_index++)
+            {
+                if(raw_version[version_index] == '.' && !dotFound)
+                {
+                    dotFound = true;
+                }
+                else if(raw_version[version_index] == '.' && dotFound)
+                {
+                    raw_version[version_index] = '\0';
+                    break;
+                }
+            }
+            result_path[strlen(result_path) -1 + strlen("/Platforms/iPhoneOS.platform/DeviceSupport/")] = '\0';
+            strncpy(result_path + strlen(result_path) - 1, "/Platforms/iPhoneOS.platform/DeviceSupport/", strlen("/Platforms/iPhoneOS.platform/DeviceSupport/"));
+
+            /* Look at all versions in DeviceSupport to see if one is prefixed with the version we need */
+            DIR *dir;
+            struct dirent *ent;
+            if ((dir = opendir (result_path)) != NULL)
+            {
+                while ((ent = readdir (dir)) != NULL)
+                {
+                    bool isMatch = true;
+                    int matchIndex = 0;
+                    while(isMatch && matchIndex < strlen(raw_version) && matchIndex < strlen(ent->d_name))
+                    {
+                        if(ent->d_name[matchIndex] != raw_version[matchIndex])
+                        {
+                            isMatch = false;
+                        }
+                        matchIndex++;
+                    }
+                    if(isMatch)
+                    {
+                        path = CFStringCreateWithFormat(NULL, NULL, CFSTR("%s%s/DeveloperDiskImage.dmg"), result_path, ent->d_name);
+                        found = true;
+                        break;
+                    }
+                }
+                closedir (dir);
+            }
+        }
     }
 
     CFRelease(version);
@@ -473,9 +530,19 @@ void handle_device(AMDeviceRef device) {
     pid_t parent = getpid();
     int pid = fork();
     if (pid == 0) {
-        system(GDB_SHELL);      // launch gdb
-        kill(parent, SIGHUP);  // "No. I am your father."
-        _exit(0);
+        FILE* select_path = popen("xcode-select --print-path", "r");
+        if (select_path)
+        {
+            char buffer[1024];
+            char* result_path = fgets(buffer, sizeof(buffer), select_path);
+            result_path[strlen(result_path) - 1] = '\0';
+            char* gdb_shell = malloc(sizeof(char)*(strlen(result_path) + strlen("/Platforms/iPhoneOS.platform/Developer/usr/libexec/gdb/gdb-arm-apple-darwin --arch armv7 -q -x ") + strlen(PREP_CMDS_PATH) + 1));
+            sprintf(gdb_shell, "%s%s%s", result_path, "/Platforms/iPhoneOS.platform/Developer/usr/libexec/gdb/gdb-arm-apple-darwin --arch armv7 -q -x ", PREP_CMDS_PATH);
+            system(gdb_shell);      // launch gdb
+            free(gdb_shell);
+            kill(parent, SIGHUP);  // "No. I am your father."
+            _exit(0);
+        }
     }
 }
